src/api/pyproject.toml
[project]
name = "kartograph-api"
version = "0.1.0"
description = "Enterprise-Ready Bi-Temporal Knowledge Graphs as a Service"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.123.9",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
    "ruff>=0.14.8",
    "types-psycopg2>=2.9.21.20251012",
]
